my app

+++ try/app1/views.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
+import requests,json
+from requests_oauthlib import OAuth1
 from django.contrib.auth.models import User
 from django.shortcuts import render
-from .forms import contactform,signupform,marksheetform
-from .models import signup,marksheet
+from .forms import contactform,signupform,marksheetform,twitterform
+from .models import signup,marksheet,twitter
 # Create your views here.
 def play(request):
 	title = 'SignUp'
@@ -49,6 +51,7 @@ def add(request):
 
 		if instance.fullname:
 		    instance.userid=request.user.id
+		    print instance.fullname
 		    instance.save() 
 		    context = {
 		"title":"MARK ADDED"
@@ -65,10 +68,13 @@ def deletes(request,pk):
 	querry = marksheet.objects.get(id=pk)
 	querry1 = querry.userid
 	try:
-		int(request.user.id) == int(querry1)
-		marksheet.objects.filter(id=pk).delete()
+		if int(request.user.id) == int(querry1):
+			marksheet.objects.filter(id=pk).delete()
 		context = {"title":"OOPS! ENTERY DELETED!!"
 		}
+
+	    
+		
 	except:
 		context={"title":"YOU CAN'T DELETE IT!!"
 		}
@@ -77,3 +83,39 @@ def deletes(request,pk):
 	
 	
 	return render(request,"delete.html",context)
+
+def tweets(request):
+	form = twitterform(request.POST or None)
+	title = "ENTER YOUR USER NAME "
+	context={
+	"form":form,
+	"title":title
+	}
+	if form.is_valid():
+		instance =form.save(commit=False)
+
+		if instance.fullname:
+			API_Key= "VBTj4JqebIFDctJlztF2xpd0P"
+			API_Secret =  "pVlDfHthTeQhsqVqVRgNduwwQIPwXxE08LKBlzPqCq2SIAWo1J"
+			Access_Token ="1010899301818658820-JpNNgTA5RMIiX7Ufe9fqUAcBgz1ora"
+			Access_Token_Secret ="GHRf6rPmX4x0zpQ6YdLD6dboe1E71d9XHqRnJN60wRqIS"
+			username = instance.fullname
+			url=" https://api.twitter.com/1.1/statuses/user_timeline.json?screen_name="+username+"&tweet_mode=extended&count=10"
+			auth = OAuth1(API_Key,API_Secret,Access_Token,Access_Token_Secret)
+			r = requests.get(url,auth=auth)
+			d= r.json()
+			a=[]
+			#print r.status_code,d
+			for i in d:
+				if i["full_text"]:
+					a.append(i["full_text"])
+					
+			instance.save()
+			print a 
+			context = {
+		"title":"YOUR LAST TWEETS ARE :",
+		"a":a
+
+		}
+
+	return render(request,"tweet.html",context)
